+++ src/redux/helpers/actions.js
@@ -1,0 +1,7 @@
+import generateRequestActions from 'redux/factories/generateRequestActions';
+
+const actions = {
+  ...generateRequestActions(['export'], 'file', 'file'),
+};
+
+export default actions;